+++ Source/CesiumRuntime/Public/CesiumPropertyTable.h
@@ -25,7 +25,7 @@ enum class ECesiumPropertyTableStatus : uint8 {
   /* The property table instance was not initialized from an actual glTF
      property table. */
   ErrorInvalidPropertyTable,
-  /* The property table's class could be found in the schema of the metadata
+  /* The property table's class could not be found in the schema of the metadata
      extension. */
   ErrorInvalidPropertyTableClass
 };
@@ -44,32 +44,32 @@ public:
    * Construct an empty property table instance.
    */
   FCesiumPropertyTable()
-      : _status(ECesiumPropertyTableStatus::ErrorInvalidPropertyTable){};
+      : _status(ECesiumPropertyTableStatus::ErrorInvalidPropertyTable) {}
 
   /**
    * Constructs a property table from a glTF Property Table.
    *
-   * @param Model The model that stores EXT_structural_metadata.
-   * @param PropertyTable The target property table.
+   * @param model The model that stores EXT_structural_metadata.
+   * @param propertyTable The target property table.
    */
   FCesiumPropertyTable(
-      const CesiumGltf::Model& Model,
-      const CesiumGltf::PropertyTable& PropertyTable)
+      const CesiumGltf::Model& model,
+      const CesiumGltf::PropertyTable& propertyTable)
       : FCesiumPropertyTable(
-            Model,
-            PropertyTable,
-            FCesiumMetadataEnumCollection::GetOrCreateFromModel(Model)) {}
+            model,
+            propertyTable,
+            FCesiumMetadataEnumCollection::GetOrCreateFromModel(model)) {}
 
   /**
    * Constructs a property table from a glTF Property Table.
    *
-   * @param Model The model that stores EXT_structural_metadata.
-   * @param PropertyTable The target property table.
+   * @param model The model that stores EXT_structural_metadata.
+   * @param propertyTable The target property table.
    * @param pEnumCollection The enum collection to use, if any.
    */
   FCesiumPropertyTable(
-      const CesiumGltf::Model& Model,
-      const CesiumGltf::PropertyTable& PropertyTable,
+      const CesiumGltf::Model& model,
+      const CesiumGltf::PropertyTable& propertyTable,
       const TSharedPtr<FCesiumMetadataEnumCollection>& pEnumCollection);
 
   /**